Updated Skype API connection
Due to recent changed with the Skype client, the Skype4COM API can no longer send messages. The Chatbot has been updated to use a third party library instead, called Skype4Sharp, which connects to the Skype REST API. All functionality with the chatbot is the same, including the response processing. A login is now required within the application, but this stops the need for Skype to be running.